rename: 패키지명 수정 (MIKKI-69)
